Fix a few compile errors

+++ terra-offheap/src/main/java/com/ritualsoftheold/terra/offheap/io/dummy/DummyOctreeLoader.java
@@ -24,7 +24,7 @@ public class DummyOctreeLoader implements OctreeLoader {
     }
 
     @Override
-    public void saveOctrees(byte index, long addr) {
+    public void saveOctrees(int index, long addr) {
         // Do nothing
     }
